Refine stream ingress normalization boundaries
