+++ adr/ADR-038-amendment-2026-04-13.md
@@ -158,7 +158,7 @@ After this amendment:
   keyed by full path (process-global, `HashMap` + u64 tick pattern from
   `context_cache.rs`).
 - `src/runtime/task_state/lazy_task_handle.rs` — test-only `LazyTaskHandle`
-  scaffolding for the lazy-load API shape used by the cold-start tests.
+  support code for the lazy-load API shape used by the cold-start tests.
 - `src/config/startup.rs` — `StartupBudget` with `VEX_MAX_STARTUP_TASK_SCANS`
   (default 200), `VEX_STARTUP_CACHE_TTL_MS`, `VEX_TRACE_STARTUP_ALLOC`.
