+++ src/runtime/json_handoff.rs
@@ -1,19 +1,11 @@
 use super::UiUpdate;
-use crate::api::stream::MAX_TOOL_CALL_INDEX;
-use crate::api::stream::chat_compat::{
-    ChatCompatChoice, ChatCompatChunk, ChatCompatDelta, ChatCompatPayload, ChatCompatToolCallDelta,
-};
-use crate::api::stream::provider::{ProviderDelta, ProviderStreamEvent};
 use crate::runtime::AssistantPhase;
-use crate::runtime::delta_accumulator::{AccumulationError, DeltaAccumulator};
+use crate::runtime::delta_accumulator::DeltaAccumulator;
 use crate::state::{StreamBlock, ToolStatus};
-use crate::turn_evidence::{
-    SummaryRecord, TurnEvidenceRecord, command_evidence_from_tool_result,
-    note_changed_files_from_tool_call,
-};
-use crate::types::{ApiUsage, ContentBlock, StreamChunkMetadata};
+use crate::turn_evidence::{SummaryRecord, TurnEvidenceRecord, command_evidence_from_tool_result};
+use crate::types::{ContentBlock, StreamChunkMetadata};
 use crate::usage::TurnTokens;
-use chrono::{DateTime, SecondsFormat, Timelike, Utc};
+use chrono::{DateTime, SecondsFormat, Utc};
 use indexmap::IndexMap;
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeSet, HashMap};
@@ -21,9 +13,13 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicU32, Ordering};
 
 mod derived;
+mod protocol_ingress;
+mod tool_calls;
 
-use self::derived::{DerivedTurnState, empty_json_object, turn_tokens_from_usage};
+use self::derived::{DerivedTurnState, turn_tokens_from_usage};
 pub use self::derived::{runtime_approval_request_event, token_usage_from_turn_tokens};
+use self::protocol_ingress::ProtocolIngressState;
+use self::tool_calls::{PendingToolCall, PendingToolCallContext};
 
 const SYNTHETIC_FINAL_TEXT_BLOCK_START: usize = 1_000_000;
 
@@ -258,111 +254,6 @@ pub fn generate_tool_call_id(counter: &AtomicU32, entropy: u16) -> ToolCallId {
     format!("tx_{}_{entropy:04x}", count)
 }
 
-fn serialize_tool_arguments(arguments: &serde_json::Value) -> String {
-    match arguments {
-        serde_json::Value::Object(map) if map.is_empty() => String::new(),
-        _ => serde_json::to_string(arguments).unwrap_or_default(),
-    }
-}
-
-#[derive(Debug, Clone)]
-struct PendingToolCall {
-    runtime_id: ToolCallId,
-    name: String,
-    arguments: serde_json::Value,
-    raw_arguments: String,
-    start_event_id: String,
-    started_at: DateTime<Utc>,
-}
-
-#[derive(Debug, Clone)]
-struct PendingToolCallContext {
-    name: String,
-    start_event_id: String,
-}
-
-#[derive(Debug, Clone, Default)]
-struct PendingChatCompatToolCall {
-    id: String,
-    name: String,
-    pending_arguments: String,
-    started: bool,
-    stopped: bool,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-enum ProviderContentBlockCompat {
-    Text {
-        text: String,
-        #[serde(default)]
-        citations: Option<Vec<serde_json::Value>>,
-    },
-    ToolUse {
-        id: String,
-        name: String,
-        #[serde(default = "empty_json_object")]
-        input: serde_json::Value,
-        #[serde(default)]
-        metadata: Option<serde_json::Value>,
-    },
-    ToolResult {
-        tool_use_id: String,
-        content: String,
-        #[serde(default)]
-        is_error: bool,
-    },
-    Thinking {
-        thinking: String,
-        signature: String,
-    },
-    ThinkingData {
-        data: String,
-    },
-    ServerToolUse {
-        id: String,
-        name: String,
-        #[serde(default = "empty_json_object")]
-        input: serde_json::Value,
-    },
-    WebSearchToolResult {
-        tool_use_id: String,
-        #[serde(default)]
-        content: serde_json::Value,
-    },
-}
-
-#[derive(Debug, Clone)]
-enum ProviderContentBlock {
-    Text {
-        text: String,
-    },
-    ToolUse {
-        id: String,
-        name: String,
-        input: serde_json::Value,
-    },
-    ToolResult,
-    Thinking {
-        thinking: String,
-    },
-    ThinkingData {
-        data: String,
-    },
-    ServerToolUse {
-        id: String,
-        name: String,
-        input: serde_json::Value,
-    },
-    WebSearchToolResult,
-}
-
-#[derive(Debug, Clone)]
-enum ProviderBlockDelta {
-    Text(String),
-    ToolArguments(String),
-}
-
 pub struct RuntimeEnvelopeNormalizer {
     task_id: String,
     turn: u32,
@@ -376,11 +267,7 @@ pub struct RuntimeEnvelopeNormalizer {
     open_final_text_block: Option<usize>,
     next_synthetic_block_index: usize,
     delta_accumulator: Option<Arc<DeltaAccumulator>>,
-    legacy_stream_turn_started: bool,
-    legacy_open_blocks: BTreeSet<usize>,
-    legacy_turn_tokens: TurnTokens,
-    chat_compat_message_started: bool,
-    chat_compat_tools: Vec<PendingChatCompatToolCall>,
+    protocol_ingress: ProtocolIngressState,
 }
 
 impl RuntimeEnvelopeNormalizer {
@@ -413,11 +300,7 @@ impl RuntimeEnvelopeNormalizer {
             open_final_text_block: None,
             next_synthetic_block_index: SYNTHETIC_FINAL_TEXT_BLOCK_START,
             delta_accumulator,
-            legacy_stream_turn_started: false,
-            legacy_open_blocks: BTreeSet::new(),
-            legacy_turn_tokens: TurnTokens::default(),
-            chat_compat_message_started: false,
-            chat_compat_tools: Vec::new(),
+            protocol_ingress: ProtocolIngressState::default(),
         }
     }
 
@@ -433,11 +316,7 @@ impl RuntimeEnvelopeNormalizer {
         self.turn_changed_files.clear();
         self.open_final_text_block = None;
         self.next_synthetic_block_index = SYNTHETIC_FINAL_TEXT_BLOCK_START;
-        self.legacy_stream_turn_started = false;
-        self.legacy_open_blocks.clear();
-        self.legacy_turn_tokens = TurnTokens::default();
-        self.chat_compat_message_started = false;
-        self.chat_compat_tools.clear();
+        self.protocol_ingress = ProtocolIngressState::default();
 
         self.next_envelope(RuntimeEvent::TurnStart { input })
     }
@@ -611,119 +490,6 @@ impl RuntimeEnvelopeNormalizer {
             .unwrap_or_default()
     }
 
-    pub(crate) fn finish_legacy_stream_turn(&mut self) -> Vec<RuntimeEnvelope> {
-        if !self.legacy_stream_turn_started {
-            return Vec::new();
-        }
-
-        let mut envelopes = Vec::new();
-        let open_blocks = std::mem::take(&mut self.legacy_open_blocks);
-        for index in open_blocks {
-            envelopes
-                .extend(self.normalize_ui_update(&UiUpdate::StreamBlockComplete { index }, None));
-        }
-
-        let usage = token_usage_from_turn_tokens(std::mem::take(&mut self.legacy_turn_tokens));
-        envelopes.extend(self.normalize_ui_update(
-            &UiUpdate::TurnComplete,
-            Some(TurnEndContext {
-                usage,
-                changed_files: Vec::new(),
-            }),
-        ));
-
-        self.legacy_stream_turn_started = false;
-        self.chat_compat_message_started = false;
-        self.chat_compat_tools.clear();
-
-        envelopes
-    }
-
-    pub(crate) fn normalize_provider_stream_event(
-        &mut self,
-        event: ProviderStreamEvent,
-    ) -> Vec<RuntimeEnvelope> {
-        match event {
-            ProviderStreamEvent::MessageStart { message } => {
-                let mut envelopes = Vec::new();
-                if let Some(metadata) = message_start_metadata(message.metadata) {
-                    envelopes.extend(self.emit_provider_server_metadata(metadata));
-                }
-                if let Some(usage) = message.usage {
-                    envelopes.extend(self.emit_provider_usage_update(usage));
-                }
-                envelopes
-            }
-            ProviderStreamEvent::ContentBlockStart {
-                index,
-                content_block,
-            } => match decode_provider_content_block(content_block) {
-                Ok(content_block) => {
-                    self.normalize_provider_content_block_start(index, content_block)
-                }
-                Err(err) => vec![self.emit_event(RuntimeEvent::Error {
-                    code: "provider_content_block_start_decode".to_string(),
-                    message: format!(
-                        "failed to decode provider content_block_start event at index {index}: {err}"
-                    ),
-                    recoverable: true,
-                })],
-            },
-            ProviderStreamEvent::ContentBlockDelta { index, delta } => provider_block_delta(&delta)
-                .map(|delta| self.normalize_provider_block_delta(index, delta))
-                .unwrap_or_default(),
-            ProviderStreamEvent::ContentBlockStop { index } => {
-                self.normalize_provider_content_block_stop(index)
-            }
-            ProviderStreamEvent::MessageDelta { delta, usage } => {
-                let mut envelopes = Vec::new();
-                if let Some(metadata) = delta.metadata {
-                    envelopes.extend(self.emit_provider_server_metadata(metadata));
-                }
-                if let Some(usage) = usage {
-                    envelopes.extend(self.emit_provider_usage_update(usage));
-                }
-                envelopes
-            }
-            ProviderStreamEvent::MessageStop
-            | ProviderStreamEvent::Ping
-            | ProviderStreamEvent::Unknown => Vec::new(),
-            ProviderStreamEvent::Error { error } => vec![self.emit_event(RuntimeEvent::Error {
-                code: error.error_type,
-                message: error.message,
-                recoverable: true,
-            })],
-        }
-    }
-
-    pub(crate) fn normalize_chat_compat_payload(
-        &mut self,
-        payload: ChatCompatPayload,
-    ) -> Vec<RuntimeEnvelope> {
-        match payload {
-            ChatCompatPayload::Done => {
-                let envelopes = self.close_chat_compat_tool_blocks();
-                self.chat_compat_message_started = false;
-                self.chat_compat_tools.clear();
-                envelopes
-            }
-            ChatCompatPayload::Chunk(chunk) => self.normalize_chat_compat_chunk(*chunk),
-        }
-    }
-
-    pub(crate) fn emit_legacy_stream_error(
-        &mut self,
-        code: String,
-        message: String,
-    ) -> RuntimeEnvelope {
-        let _ = self.ensure_legacy_stream_turn_started();
-        self.emit_event(RuntimeEvent::Error {
-            code,
-            message,
-            recoverable: true,
-        })
-    }
-
     pub fn emit_error(
         &mut self,
         code: String,
@@ -782,271 +548,6 @@ impl RuntimeEnvelopeNormalizer {
 
     pub fn emit_event(&mut self, event: RuntimeEvent) -> RuntimeEnvelope {
         self.next_envelope(event)
-    }
-
-    fn ensure_legacy_stream_turn_started(&mut self) -> Vec<RuntimeEnvelope> {
-        if self.legacy_stream_turn_started {
-            return Vec::new();
-        }
-
-        let envelope = self.start_turn(1, None);
-        self.legacy_stream_turn_started = true;
-        vec![envelope]
-    }
-
-    fn normalize_provider_content_block_start(
-        &mut self,
-        index: usize,
-        content_block: ProviderContentBlock,
-    ) -> Vec<RuntimeEnvelope> {
-        let Some((block, initial_delta)) = provider_stream_block(&content_block) else {
-            return Vec::new();
-        };
-
-        let mut envelopes = self.ensure_legacy_stream_turn_started();
-        self.legacy_open_blocks.insert(index);
-        envelopes
-            .extend(self.normalize_ui_update(&UiUpdate::StreamBlockStart { index, block }, None));
-        if let Some(delta) = initial_delta.filter(|delta| !delta.is_empty()) {
-            envelopes.extend(
-                self.normalize_ui_update(&UiUpdate::StreamBlockDelta { index, delta }, None),
-            );
-        }
-        envelopes
-    }
-
-    fn normalize_provider_block_delta(
-        &mut self,
-        index: usize,
-        block_delta: ProviderBlockDelta,
-    ) -> Vec<RuntimeEnvelope> {
-        let mut envelopes = self.ensure_legacy_stream_turn_started();
-        if matches!(block_delta, ProviderBlockDelta::Text(_))
-            && !self.legacy_open_blocks.contains(&index)
-        {
-            self.legacy_open_blocks.insert(index);
-            envelopes.extend(self.normalize_ui_update(
-                &UiUpdate::StreamBlockStart {
-                    index,
-                    block: StreamBlock::Thinking {
-                        content: String::new(),
-                        collapsed: false,
-                    },
-                },
-                None,
-            ));
-        }
-
-        let delta = match block_delta {
-            ProviderBlockDelta::Text(delta) | ProviderBlockDelta::ToolArguments(delta) => delta,
-        };
-        envelopes
-            .extend(self.normalize_ui_update(&UiUpdate::StreamBlockDelta { index, delta }, None));
-        envelopes
-    }
-
-    fn normalize_provider_content_block_stop(&mut self, index: usize) -> Vec<RuntimeEnvelope> {
-        if !self.legacy_open_blocks.remove(&index) {
-            return Vec::new();
-        }
-
-        self.normalize_ui_update(&UiUpdate::StreamBlockComplete { index }, None)
-    }
-
-    fn emit_provider_server_metadata(
-        &mut self,
-        metadata: StreamChunkMetadata,
-    ) -> Vec<RuntimeEnvelope> {
-        let mut envelopes = self.ensure_legacy_stream_turn_started();
-        envelopes
-            .extend(self.normalize_ui_update(&UiUpdate::ServerMetadata(Box::new(metadata)), None));
-        envelopes
-    }
-
-    fn emit_provider_usage_update(&mut self, usage: ApiUsage) -> Vec<RuntimeEnvelope> {
-        accumulate_turn_tokens(&mut self.legacy_turn_tokens, &usage);
-
-        let mut envelopes = self.ensure_legacy_stream_turn_started();
-        if let Some(usage) = token_usage_from_api_usage(&usage) {
-            envelopes.push(self.emit_event(RuntimeEvent::UsageUpdated { usage }));
-        }
-        envelopes
-    }
-
-    fn normalize_chat_compat_chunk(&mut self, mut chunk: ChatCompatChunk) -> Vec<RuntimeEnvelope> {
-        let mut envelopes = self.emit_chat_compat_message_start(&chunk);
-
-        if let Some(mut usage) = chunk.usage.take() {
-            if usage.service_tier.is_none() {
-                usage.service_tier = chunk.service_tier.clone();
-            }
-            if let Some(metadata) = chat_compat_metadata(&chunk, None, None) {
-                envelopes.extend(self.emit_provider_server_metadata(metadata));
-            }
-            envelopes.extend(self.emit_provider_usage_update(usage));
-        }
-
-        if chunk.choices.is_empty() {
-            return envelopes;
-        }
-
-        for choice in std::mem::take(&mut chunk.choices) {
-            let ChatCompatChoice {
-                index: choice_index,
-                delta,
-                finish_reason,
-                logprobs,
-            } = choice;
-            let ChatCompatDelta {
-                role: _,
-                content,
-                refusal,
-                tool_calls,
-            } = delta;
-
-            if let Some(content) = content {
-                envelopes.extend(
-                    self.normalize_provider_block_delta(0, ProviderBlockDelta::Text(content)),
-                );
-            }
-
-            if let Some(tool_calls) = tool_calls {
-                for tool_call in tool_calls {
-                    envelopes.extend(self.apply_chat_compat_tool_delta(choice_index, tool_call));
-                }
-            }
-
-            let has_logprobs = logprobs.is_some();
-            let metadata = chat_compat_metadata(&chunk, choice_index, logprobs);
-            let has_server_progress = metadata
-                .as_ref()
-                .is_some_and(|md| md.prompt_progress.is_some() || md.timings.is_some());
-
-            if (refusal.is_some() || finish_reason.is_some() || has_logprobs || has_server_progress)
-                && let Some(metadata) = metadata
-            {
-                envelopes.extend(self.emit_provider_server_metadata(metadata));
-            }
-
-            if finish_reason.is_some() {
-                envelopes.extend(self.close_chat_compat_tool_blocks());
-            }
-        }
-
-        envelopes
-    }
-
-    fn emit_chat_compat_message_start(&mut self, chunk: &ChatCompatChunk) -> Vec<RuntimeEnvelope> {
-        if self.chat_compat_message_started {
-            return Vec::new();
-        }
-
-        let role = chunk
-            .choices
-            .iter()
-            .find_map(|choice| choice.delta.role.clone());
-        let metadata = chat_compat_metadata(chunk, None, None);
-        let has_message_fields =
-            chunk.id.is_some() || chunk.model.is_some() || role.is_some() || metadata.is_some();
-        if !has_message_fields {
-            return Vec::new();
-        }
-
-        self.chat_compat_message_started = true;
-        message_start_metadata(metadata)
-            .map(|metadata| self.emit_provider_server_metadata(metadata))
-            .unwrap_or_default()
-    }
-
-    fn apply_chat_compat_tool_delta(
-        &mut self,
-        choice_index: Option<usize>,
-        tool_call: ChatCompatToolCallDelta,
-    ) -> Vec<RuntimeEnvelope> {
-        let raw_index = tool_call.index.unwrap_or(0).min(MAX_TOOL_CALL_INDEX);
-        let block_index = raw_index.saturating_add(1);
-        self.ensure_chat_compat_tool_state(block_index);
-
-        let mut start_block = None;
-        let mut partial_json = None;
-        {
-            let state = &mut self.chat_compat_tools[block_index];
-            let _ = (choice_index, tool_call.call_type);
-
-            if let Some(id) = tool_call.id
-                && !id.is_empty()
-            {
-                state.id = id;
-            }
-            if let Some(function) = tool_call.function {
-                if let Some(name) = function.name
-                    && !name.is_empty()
-                {
-                    state.name = name;
-                }
-                if let Some(arguments) = function.arguments {
-                    state.pending_arguments.push_str(&arguments);
-                }
-            }
-
-            if !state.started && !state.name.is_empty() {
-                let id = if state.id.is_empty() {
-                    format!("toolu_chat_compat_{block_index}")
-                } else {
-                    state.id.clone()
-                };
-                start_block = Some(ProviderContentBlock::ToolUse {
-                    id,
-                    name: state.name.clone(),
-                    input: empty_json_object(),
-                });
-                state.started = true;
-            }
-
-            if state.started && !state.pending_arguments.is_empty() {
-                partial_json = Some(std::mem::take(&mut state.pending_arguments));
-            }
-        }
-
-        let mut envelopes = Vec::new();
-        if let Some(content_block) = start_block {
-            envelopes
-                .extend(self.normalize_provider_content_block_start(block_index, content_block));
-        }
-        if let Some(partial_json) = partial_json {
-            envelopes.extend(self.normalize_provider_block_delta(
-                block_index,
-                ProviderBlockDelta::ToolArguments(partial_json),
-            ));
-        }
-        envelopes
-    }
-
-    fn ensure_chat_compat_tool_state(&mut self, index: usize) {
-        let required = index.saturating_add(1);
-        if self.chat_compat_tools.len() < required {
-            self.chat_compat_tools
-                .resize_with(required, PendingChatCompatToolCall::default);
-        }
-    }
-
-    fn close_chat_compat_tool_blocks(&mut self) -> Vec<RuntimeEnvelope> {
-        let mut to_close = Vec::new();
-        for (index, state) in self.chat_compat_tools.iter_mut().enumerate() {
-            if index == 0 {
-                continue;
-            }
-            if state.started && !state.stopped {
-                state.stopped = true;
-                to_close.push(index);
-            }
-        }
-
-        let mut envelopes = Vec::new();
-        for index in to_close {
-            envelopes.extend(self.normalize_provider_content_block_stop(index));
-        }
-        envelopes
     }
 
     fn complete_turn(&mut self, turn_end: TurnEndContext) -> Vec<RuntimeEnvelope> {
@@ -1115,288 +616,6 @@ impl RuntimeEnvelopeNormalizer {
         )]
     }
 
-    fn normalize_grammar_tool_call(
-        &mut self,
-        tool_call_value: &serde_json::Value,
-    ) -> Option<RuntimeEnvelope> {
-        let name = tool_call_value.get("name")?.as_str()?.to_string();
-        let arguments = tool_call_value
-            .get("arguments")
-            .cloned()
-            .unwrap_or_else(empty_json_object);
-
-        let runtime_id = self.generate_tool_call_id();
-        let started_at = Utc::now();
-        let envelope = self.next_envelope(RuntimeEvent::ToolCallStarted {
-            tool_call_id: runtime_id.clone(),
-            tool_name: name.clone(),
-            arguments: arguments.clone(),
-            status: ToolStatus::Pending,
-            started_at: timestamp_string(started_at),
-        });
-        self.pending_tool_calls.insert(
-            runtime_id.clone(),
-            PendingToolCall {
-                runtime_id: runtime_id.clone(),
-                name: name.clone(),
-                arguments: arguments.clone(),
-                raw_arguments: serialize_tool_arguments(&arguments),
-                start_event_id: envelope.event_id.clone(),
-                started_at,
-            },
-        );
-        self.pending_tool_call_contexts.insert(
-            runtime_id.clone(),
-            PendingToolCallContext {
-                name: name.clone(),
-                start_event_id: envelope.event_id.clone(),
-            },
-        );
-        self.start_tool_accumulator(&runtime_id, &name);
-        self.seed_tool_accumulator(&runtime_id, &arguments);
-
-        Some(envelope)
-    }
-
-    fn record_tool_call(
-        &mut self,
-        source_id: String,
-        name: String,
-        arguments: serde_json::Value,
-    ) -> RuntimeEnvelope {
-        let runtime_id = self.generate_tool_call_id();
-        let started_at = Utc::now();
-        let envelope = self.next_envelope(RuntimeEvent::ToolCallStarted {
-            tool_call_id: runtime_id.clone(),
-            tool_name: name.clone(),
-            arguments: arguments.clone(),
-            status: ToolStatus::Pending,
-            started_at: timestamp_string(started_at),
-        });
-        self.pending_tool_calls.insert(
-            source_id,
-            PendingToolCall {
-                runtime_id: runtime_id.clone(),
-                name: name.clone(),
-                arguments: arguments.clone(),
-                raw_arguments: serialize_tool_arguments(&arguments),
-                start_event_id: envelope.event_id.clone(),
-                started_at,
-            },
-        );
-        self.pending_tool_call_contexts.insert(
-            runtime_id.clone(),
-            PendingToolCallContext {
-                name: name.clone(),
-                start_event_id: envelope.event_id.clone(),
-            },
-        );
-        self.start_tool_accumulator(&runtime_id, &name);
-        self.seed_tool_accumulator(&runtime_id, &arguments);
-
-        envelope
-    }
-
-    fn record_tool_result(
-        &mut self,
-        source_id: &str,
-        output: String,
-        is_error: bool,
-    ) -> RuntimeEnvelope {
-        let completed_at = Utc::now();
-        let completed_at_value = timestamp_string(completed_at);
-        if let Some(pending) = self.pending_tool_calls.shift_remove(source_id) {
-            self.pending_tool_call_contexts.remove(&pending.runtime_id);
-            self.finish_tool_accumulator(&pending.runtime_id);
-            if !is_error {
-                note_changed_files_from_tool_call(
-                    &mut self.turn_changed_files,
-                    &pending.name,
-                    &pending.arguments,
-                );
-            }
-            let duration_ms = completed_at
-                .signed_duration_since(pending.started_at)
-                .num_milliseconds()
-                .max(0) as u64;
-            let event = if is_error {
-                RuntimeEvent::ToolCallFailed {
-                    tool_call_id: pending.runtime_id,
-                    tool_name: Some(pending.name),
-                    status: ToolStatus::Error,
-                    started_at: Some(timestamp_string(pending.started_at)),
-                    completed_at: completed_at_value,
-                    duration_ms: Some(duration_ms),
-                    output,
-                }
-            } else {
-                RuntimeEvent::ToolCallCompleted {
-                    tool_call_id: pending.runtime_id,
-                    tool_name: Some(pending.name),
-                    status: ToolStatus::Complete,
-                    started_at: Some(timestamp_string(pending.started_at)),
-                    completed_at: completed_at_value,
-                    duration_ms: Some(duration_ms),
-                    output,
-                }
-            };
-            return self.next_envelope_with_context(event, None, Some(pending.start_event_id));
-        }
-
-        self.next_envelope(if is_error {
-            RuntimeEvent::ToolCallFailed {
-                tool_call_id: source_id.to_string(),
-                tool_name: None,
-                status: ToolStatus::Error,
-                started_at: None,
-                completed_at: completed_at_value,
-                duration_ms: None,
-                output,
-            }
-        } else {
-            RuntimeEvent::ToolCallCompleted {
-                tool_call_id: source_id.to_string(),
-                tool_name: None,
-                status: ToolStatus::Complete,
-                started_at: None,
-                completed_at: completed_at_value,
-                duration_ms: None,
-                output,
-            }
-        })
-    }
-
-    fn record_tool_call_arguments_delta(
-        &mut self,
-        tool_call_id: &ToolCallId,
-        delta: &str,
-    ) -> RuntimeEnvelope {
-        let (tool_name, parent_event_id) = self
-            .pending_tool_call_contexts
-            .get(tool_call_id)
-            .map(|pending| {
-                (
-                    Some(pending.name.clone()),
-                    Some(pending.start_event_id.clone()),
-                )
-            })
-            .unwrap_or((None, None));
-        let invalid_json = matches!(
-            self.accumulate_tool_delta(tool_call_id, delta),
-            Some(AccumulationError::MalformedPartial)
-        );
-        let arguments = self.append_pending_tool_delta(tool_call_id, delta);
-
-        self.next_envelope_with_context(
-            RuntimeEvent::ToolCallArgumentsDelta {
-                tool_call_id: tool_call_id.clone(),
-                tool_name,
-                delta: delta.to_string(),
-                status: ToolStatus::Pending,
-                arguments,
-                invalid_json,
-            },
-            None,
-            parent_event_id,
-        )
-    }
-
-    fn resolve_changed_files(&self, changed_files: Vec<String>) -> Vec<String> {
-        if changed_files.is_empty() {
-            self.turn_changed_files.iter().cloned().collect()
-        } else {
-            changed_files
-        }
-    }
-
-    fn start_tool_accumulator(&self, tool_call_id: &ToolCallId, name: &str) {
-        let Some(delta_accumulator) = &self.delta_accumulator else {
-            return;
-        };
-
-        if let Err(error) = delta_accumulator.start_tool(
-            tool_call_id.clone(),
-            self.task_id.clone(),
-            name.to_string(),
-        ) {
-            tracing::debug!(%error, tool_call_id = %tool_call_id, "failed to start tool delta accumulation");
-        }
-    }
-
-    fn seed_tool_accumulator(&self, tool_call_id: &ToolCallId, arguments: &serde_json::Value) {
-        let Some(delta_accumulator) = &self.delta_accumulator else {
-            return;
-        };
-
-        let Ok(serialized) = serde_json::to_string(arguments) else {
-            return;
-        };
-        if serialized == "{}" {
-            return;
-        }
-
-        if let Err(error) = delta_accumulator.accumulate(tool_call_id, &serialized) {
-            tracing::debug!(%error, tool_call_id = %tool_call_id, "failed to seed tool delta accumulation");
-        }
-    }
-
-    fn accumulate_tool_delta(
-        &self,
-        tool_call_id: &ToolCallId,
-        delta: &str,
-    ) -> Option<AccumulationError> {
-        let Some(delta_accumulator) = &self.delta_accumulator else {
-            return None;
-        };
-
-        if let Err(error) = delta_accumulator.accumulate(tool_call_id, delta) {
-            tracing::debug!(%error, tool_call_id = %tool_call_id, "failed to accumulate tool delta");
-            return Some(error);
-        }
-
-        None
-    }
-
-    fn finish_tool_accumulator(&self, tool_call_id: &ToolCallId) {
-        let Some(delta_accumulator) = &self.delta_accumulator else {
-            return;
-        };
-
-        delta_accumulator.finish(tool_call_id);
-    }
-
-    fn append_pending_tool_delta(
-        &mut self,
-        tool_call_id: &ToolCallId,
-        delta: &str,
-    ) -> Option<serde_json::Value> {
-        let pending = self
-            .pending_tool_calls
-            .values_mut()
-            .find(|pending| pending.runtime_id == *tool_call_id)?;
-
-        pending.raw_arguments.push_str(delta);
-        let arguments = serde_json::from_str::<serde_json::Value>(&pending.raw_arguments)
-            .or_else(|_| serde_json::from_str(pending.raw_arguments.trim()))
-            .ok()?;
-        pending.arguments = arguments.clone();
-        Some(arguments)
-    }
-
-    fn finish_pending_tool_accumulations(&self) {
-        for pending in self.pending_tool_calls.values() {
-            self.finish_tool_accumulator(&pending.runtime_id);
-        }
-    }
-
-    fn generate_tool_call_id(&mut self) -> ToolCallId {
-        let now = Utc::now();
-        let entropy = ((now.nanosecond() as u128
-            ^ self.tool_id_counter.load(Ordering::SeqCst) as u128)
-            & 0xffff) as u16;
-        generate_tool_call_id(&self.tool_id_counter, entropy)
-    }
-
     fn next_envelope(&mut self, event: RuntimeEvent) -> RuntimeEnvelope {
         self.next_envelope_with_context(event, None, None)
     }
@@ -1432,198 +651,6 @@ impl RuntimeEnvelopeNormalizer {
         };
         self.next_seq = seq.saturating_add(1);
         envelope
-    }
-}
-
-fn decode_provider_content_block(value: serde_json::Value) -> Result<ProviderContentBlock, String> {
-    let block_type = value
-        .get("type")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<missing>")
-        .to_string();
-
-    serde_json::from_value::<ProviderContentBlockCompat>(value)
-        .map(Into::into)
-        .map_err(|err| format!("type={block_type}: {err}"))
-}
-
-impl From<ProviderContentBlockCompat> for ProviderContentBlock {
-    fn from(value: ProviderContentBlockCompat) -> Self {
-        match value {
-            ProviderContentBlockCompat::Text { text, citations } => {
-                let _ = citations;
-                Self::Text { text }
-            }
-            ProviderContentBlockCompat::ToolUse {
-                id,
-                name,
-                input,
-                metadata,
-            } => {
-                let _ = metadata;
-                Self::ToolUse { id, name, input }
-            }
-            ProviderContentBlockCompat::ToolResult {
-                tool_use_id,
-                content,
-                is_error,
-            } => {
-                let _ = (tool_use_id, content, is_error);
-                Self::ToolResult
-            }
-            ProviderContentBlockCompat::Thinking {
-                thinking,
-                signature,
-            } => {
-                let _ = signature;
-                Self::Thinking { thinking }
-            }
-            ProviderContentBlockCompat::ThinkingData { data } => Self::ThinkingData { data },
-            ProviderContentBlockCompat::ServerToolUse { id, name, input } => {
-                Self::ServerToolUse { id, name, input }
-            }
-            ProviderContentBlockCompat::WebSearchToolResult {
-                tool_use_id,
-                content,
-            } => {
-                let _ = (tool_use_id, content);
-                Self::WebSearchToolResult
-            }
-        }
-    }
-}
-
-fn provider_stream_block(
-    content_block: &ProviderContentBlock,
-) -> Option<(StreamBlock, Option<String>)> {
-    match content_block {
-        ProviderContentBlock::Text { text } => Some((
-            StreamBlock::Thinking {
-                content: String::new(),
-                collapsed: false,
-            },
-            Some(text.clone()),
-        )),
-        ProviderContentBlock::Thinking { thinking } => Some((
-            StreamBlock::Thinking {
-                content: String::new(),
-                collapsed: false,
-            },
-            Some(thinking.clone()),
-        )),
-        ProviderContentBlock::ThinkingData { data } => Some((
-            StreamBlock::Thinking {
-                content: String::new(),
-                collapsed: false,
-            },
-            Some(data.clone()),
-        )),
-        ProviderContentBlock::ToolUse { id, name, input }
-        | ProviderContentBlock::ServerToolUse { id, name, input } => Some((
-            StreamBlock::ToolCall {
-                id: id.clone(),
-                name: name.clone(),
-                input: input.clone(),
-                status: ToolStatus::Pending,
-            },
-            None,
-        )),
-        ProviderContentBlock::ToolResult | ProviderContentBlock::WebSearchToolResult => None,
-    }
-}
-
-fn provider_block_delta(delta: &ProviderDelta) -> Option<ProviderBlockDelta> {
-    if let Some(text) = delta.text.as_deref()
-        && !text.is_empty()
-    {
-        return Some(ProviderBlockDelta::Text(text.to_string()));
-    }
-    if let Some(thinking) = delta.thinking.as_deref()
-        && !thinking.is_empty()
-    {
-        return Some(ProviderBlockDelta::Text(thinking.to_string()));
-    }
-    if let Some(partial_json) = delta.partial_json.as_deref()
-        && !partial_json.is_empty()
-    {
-        return Some(ProviderBlockDelta::ToolArguments(partial_json.to_string()));
-    }
-
-    None
-}
-
-fn message_start_metadata(metadata: Option<StreamChunkMetadata>) -> Option<StreamChunkMetadata> {
-    let mut metadata = metadata?;
-    metadata.prompt_progress = None;
-    metadata.timings = None;
-
-    (metadata.object.is_some()
-        || metadata.created.is_some()
-        || metadata.system_fingerprint.is_some()
-        || metadata.service_tier.is_some()
-        || metadata.choice_index.is_some()
-        || metadata.logprobs.is_some())
-    .then_some(metadata)
-}
-
-fn chat_compat_metadata(
-    chunk: &ChatCompatChunk,
-    choice_index: Option<usize>,
-    logprobs: Option<serde_json::Value>,
-) -> Option<StreamChunkMetadata> {
-    let metadata = StreamChunkMetadata {
-        object: chunk.object.clone(),
-        created: chunk.created,
-        system_fingerprint: chunk.system_fingerprint.clone(),
-        service_tier: chunk.service_tier.clone(),
-        choice_index,
-        logprobs,
-        prompt_progress: chunk.prompt_progress.clone(),
-        timings: chunk.timings.clone(),
-    };
-    (metadata.object.is_some()
-        || metadata.created.is_some()
-        || metadata.system_fingerprint.is_some()
-        || metadata.service_tier.is_some()
-        || metadata.choice_index.is_some()
-        || metadata.logprobs.is_some()
-        || metadata.prompt_progress.is_some()
-        || metadata.timings.is_some())
-    .then_some(metadata)
-}
-
-fn token_usage_from_api_usage(usage: &ApiUsage) -> Option<TokenUsageEnvelope> {
-    let usage = TokenUsageEnvelope {
-        input: usage.input_tokens.unwrap_or(0),
-        output: usage.output_tokens.unwrap_or(0),
-        estimated: false,
-        cache_creation_input: usage.cache_creation_input_tokens.unwrap_or(0),
-        cache_read_input: usage.cache_read_input_tokens.unwrap_or(0),
-    };
-
-    (!usage.input.eq(&0)
-        || !usage.output.eq(&0)
-        || !usage.cache_creation_input.eq(&0)
-        || !usage.cache_read_input.eq(&0))
-    .then_some(usage)
-}
-
-fn accumulate_turn_tokens(turn_tokens: &mut TurnTokens, usage: &ApiUsage) {
-    if let Some(input) = usage.input_tokens {
-        turn_tokens.input = turn_tokens.input.saturating_add(input);
-    }
-    if let Some(output) = usage.output_tokens {
-        turn_tokens.output = turn_tokens.output.saturating_add(output);
-    }
-    if let Some(cache_creation) = usage.cache_creation_input_tokens {
-        turn_tokens.cache_creation_input_tokens = turn_tokens
-            .cache_creation_input_tokens
-            .saturating_add(cache_creation);
-    }
-    if let Some(cache_read) = usage.cache_read_input_tokens {
-        turn_tokens.cache_read_input_tokens = turn_tokens
-            .cache_read_input_tokens
-            .saturating_add(cache_read);
     }
 }
